Refactor make the dropdown menu reusable & change locale folder path
