+++ src/types/props.ts
@@ -1,6 +1,9 @@
-export interface FilterType {
-  setRegion: (name: string) => void;
+export interface DropDownType {
+  handleChange: (value: string) => void;
   loading: boolean;
+  placeholder: string;
+  className: string;
+  options: { label: string; value: string }[];
 }
 export interface SearchBarType {
   setName: (name: string) => void;
